Split cdcacm.c
Split up cdcacm.c so that it only contains USB code.  main() and all
baseband processing code is moved to main.c.

The code itself has been changed only minorly, to avoid confusion, and
could still use some more cleanup.  The input ring buffer is passed to
the USB processing code through cdcacm_init().
