+++ include/cdcacm.h
@@ -1,0 +1,8 @@
+#ifndef __CDCACM_H
+#define __CDCACM_H
+
+#include "ring.h"
+
+void cdcacm_init(struct ring *rx);
+
+#endif /* __CDCACM_H */
